Remove type property from @Checksum annotation
The type property in the @Checksum annotation has been removed to streamline the annotation setup. Instead of manually specifying the type, it is now retrieved directly from the interface method's return type. This change is part of an effort to simplify checksum verification and improve code readability.

+++ src/main/java/io/github/mtrevisan/boxon/annotations/checksummers/CRC16CCITT.java
@@ -47,8 +47,8 @@ public final class CRC16CCITT implements Checksummer{
 
 
 	@Override
-	public short calculateChecksum(final byte[] data, final int start, final int end, final int startValue){
-		short value = (short)startValue;
+	public short calculateChecksum(final byte[] data, final int start, final int end, final short startValue){
+		short value = startValue;
 		for(int i = Math.max(start, 0), length = Math.min(end, data.length); i < length; i ++){
 			final byte datum = data[i];
