+++ src/main/java/io/github/mtrevisan/boxon/annotations/checksummers/Checksummer.java
@@ -37,6 +37,6 @@ public interface Checksummer{
 	 * @param startValue	The starting value of the checksum.
 	 * @return	The checksum.
 	 */
-	short calculateChecksum(byte[] data, int start, int end, int startValue);
+	short calculateChecksum(byte[] data, int start, int end, short startValue);
 
 }
